try to find way how is works

+++ src/main/java/com/httpservlet/test/httpservlets/servlet/utils/dao/FlightDao.java
@@ -48,7 +48,7 @@ public class FlightDao implements Dao<Long, Flight> {
             res.getObject("departure_airport_code", String.class),
             res.getObject("arrival_date", Timestamp.class).toLocalDateTime(),
             res.getObject("arrival_airport_code", String.class),
-            res.getObject("aircraft_id", Long.class),
+            res.getObject("aircraft_id", Integer.class),
             FlightStatus.valueOf(res.getObject("status", String.class))
         );
     }
